@@ -6,9 +6,9 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 class FlowLogDesc {
-  final String detail;
-  final bool hasError;
-  final bool hasItem;
+  final String? detail;
+  final bool? hasError;
+  final bool? hasItem;
 
   FlowLogDesc({
     this.detail,
@@ -18,29 +18,29 @@ class FlowLogDesc {
 
   @override
   String toString() {
-    return detail;
+    return detail!;
   }
 }
 
 // 一个工作流
 class FlowLog {
-  final String name;
-  List<Log> logs;
+  final String? name;
+  List<Log>? logs;
 
-  DateTime createdAt;
+  DateTime? createdAt;
 
   /// 超时时间，如果超过这个时间就算作一个新Flow
-  final Duration timeout;
+  final Duration? timeout;
 
-  DateTime _endAt;
-  DateTime get endAt => _endAt;
+  DateTime? _endAt;
+  DateTime? get endAt => _endAt;
 
   FlowLog._({
     this.name,
     this.logs,
     this.timeout,
     this.createdAt,
-    DateTime end,
+    DateTime? end,
   }) : _endAt = end {
     logs ??= [];
   }
@@ -48,49 +48,49 @@ class FlowLog {
   FlowLog({
     this.name: "system",
     this.logs,
-    Duration timeout: const Duration(seconds: 30),
+    Duration? timeout: const Duration(seconds: 30),
   })  : createdAt = DateTime.now(),
         this.timeout = timeout ?? const Duration(seconds: 30) {
     logs ??= [];
-    FlowCenter.instance.workingFlow[name] = this;
+    FlowCenter.instance!.workingFlow[name] = this;
   }
 
-  DateTime get latestTime {
-    if (logs.length == 0) {
+  DateTime? get latestTime {
+    if (logs!.length == 0) {
       return createdAt;
     }
-    return logs.last.dateTime;
+    return logs!.last.dateTime;
   }
 
   DateTime get expireTime =>
-      latestTime.add(timeout ?? const Duration(seconds: 30));
+      latestTime!.add(timeout ?? const Duration(seconds: 30));
 
   bool get isTimeout =>
       expireTime.millisecondsSinceEpoch < DateTime.now().millisecondsSinceEpoch;
 
-  Timer timer;
+  Timer? timer;
 
   /// 增加一个新log，如果超时了，就总结log并加入已完成
   _addRawLog(Log log) {
     if (isTimeout) {
       this.end(
-        'End with timeout(${timeout.inSeconds}s)',
+        'End with timeout(${timeout!.inSeconds}s)',
         LogType.error,
       );
     } else {
       timer?.cancel();
       timer = null;
-      timer = Timer(timeout, () {
+      timer = Timer(timeout!, () {
         this.end(
-          'End with timeout(${timeout.inSeconds}s)',
+          'End with timeout(${timeout!.inSeconds}s)',
           LogType.error,
         );
         timer?.cancel();
         timer = null;
       });
     }
-    this.logs.add(log);
-    FlowCenter.instance._notify();
+    this.logs!.add(log);
+    FlowCenter.instance!._notify();
   }
 
   /// 添加一个常规log
@@ -109,32 +109,32 @@ class FlowLog {
   }
 
   /// 结束当前Flow
-  end([dynamic log, LogType type]) {
+  end([dynamic log, LogType? type]) {
     if (log != null) {
-      this.logs.add(Log(log, type ?? LogType.log));
+      this.logs!.add(Log(log, type ?? LogType.log));
     }
     timer?.cancel();
     timer = null;
     _endAt = DateTime.now();
     // 拷贝一份去
-    FlowCenter.instance.flowList.add(FlowLog._(
+    FlowCenter.instance!.flowList.add(FlowLog._(
       name: name,
-      logs: List.from(logs),
+      logs: List.from(logs!),
       timeout: timeout,
       createdAt: createdAt,
       end: endAt,
     ));
     this.createdAt = DateTime.now();
     this._endAt = null;
-    FlowCenter.instance.workingFlow.remove(name);
-    this.logs.clear();
-    FlowCenter.instance._notify();
+    FlowCenter.instance!.workingFlow.remove(name);
+    this.logs!.clear();
+    FlowCenter.instance!._notify();
   }
 
   FlowLogDesc get desc {
     int normalCount = 0;
     int errorCount = 0;
-    for (var log in logs) {
+    for (var log in logs!) {
       if (log.type == LogType.log) {
         normalCount += 1;
       } else {
@@ -144,7 +144,7 @@ class FlowLog {
     var detail = '$normalCount Logs, $errorCount Errors';
     var duration = '';
     if (endAt != null) {
-      duration = ", ${createdAt.difference(endAt).inMilliseconds * -1}ms";
+      duration = ", ${createdAt!.difference(endAt!).inMilliseconds * -1}ms";
     }
     return FlowLogDesc(
       detail: detail + duration,
@@ -154,14 +154,14 @@ class FlowLog {
   }
 
   String get startTimeText =>
-      DateFormat(FConsole.instance.options.timeFormat).format(createdAt);
+      DateFormat(FConsole.instance!.options.timeFormat).format(createdAt!);
 
   String get endTimeText => endAt == null
       ? 'null'
-      : DateFormat(FConsole.instance.options.timeFormat).format(endAt);
+      : DateFormat(FConsole.instance!.options.timeFormat).format(endAt!);
 
   String get shareText => [
-        if (logs.isNotEmpty) "Time: ${logs.first.dateTime}",
+        if (logs!.isNotEmpty) "Time: ${logs!.first.dateTime}",
         "Event: $name",
         "Overview: $desc",
         "-----------------",
@@ -171,28 +171,28 @@ class FlowLog {
 
   String get logsDesc {
     List<String> logStr = [];
-    for (var i = 0; i < logs.length; i++) {
-      var log = logs[i];
+    for (var i = 0; i < logs!.length; i++) {
+      var log = logs![i];
       if (i == 0) {
         logStr.add('[start]$log');
         continue;
       }
-      var lastLog = logs[i - 1];
-      var diff = log.dateTime.difference(lastLog.dateTime).inMilliseconds;
+      var lastLog = logs![i - 1];
+      var diff = log.dateTime!.difference(lastLog.dateTime!).inMilliseconds;
       logStr.add('[${diff}ms]$log');
     }
     return logStr.join('\n');
   }
 
   /// 通过id获取一个正在进行的Flow，如果flow还没有创建，那么就创建一个新的再返回
-  static FlowLog of(String name, [Duration initTimeOut]) {
-    if (FlowCenter.instance.workingFlow[name] == null) {
-      FlowCenter.instance.workingFlow[name] = FlowLog(
+  static FlowLog? of(String name, [Duration? initTimeOut]) {
+    if (FlowCenter.instance!.workingFlow[name] == null) {
+      FlowCenter.instance!.workingFlow[name] = FlowLog(
         name: name,
         timeout: initTimeOut,
       );
     }
-    return FlowCenter.instance.workingFlow[name];
+    return FlowCenter.instance!.workingFlow[name];
   }
 
   @override
@@ -203,7 +203,7 @@ class FlowLog {
 
 class FlowCenter extends ChangeNotifier {
   /// 正在进行的流程记录
-  Map<String, FlowLog> workingFlow = {};
+  Map<String?, FlowLog> workingFlow = {};
 
   /// 已经完成的流程记录
   List<FlowLog> flowList = [];
@@ -218,13 +218,13 @@ class FlowCenter extends ChangeNotifier {
   }
 
   // 工厂模式
-  factory FlowCenter() => _getInstance();
-  static FlowCenter get instance => _getInstance();
-  static FlowCenter _instance;
+  factory FlowCenter() => _getInstance()!;
+  static FlowCenter? get instance => _getInstance();
+  static FlowCenter? _instance;
   FlowCenter._internal() {
     // 初始化
   }
-  static FlowCenter _getInstance() {
+  static FlowCenter? _getInstance() {
     if (_instance == null) {
       _instance = FlowCenter._internal();
     }
